requirements.txt
streamlit>=1.40.0
pandas>=2.2.2
pypdf>=5.1.0
google-genai>=1.0.0
supabase>=2.10.0
python-dotenv>=1.0.0
